+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## master
 
 - Moved `-[NSString UUIDString]` to `ESUUIDString()` helper
+- Moved `-[NSString randomStringWithLength:]` to `ESRandomString()` helper
 
 ## 3.16.1 (2019-07-25)